Tests/LibWeb/Text/input/wpt-import/encoding-detection/ar-windows-1256.tentative.html

<!doctype html>
<title>ar windows-1256</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>衵ﾇ 褂 ﾇﾎﾊﾈﾇﾑ ﾊﾑ肬ﾒ ﾇ眦ﾍﾑﾝ.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1256", 'Expected windows-1256');
    }, "Check detection result");
      done();
};
</script>
